bug fix arduino

+++ src/arduino-sketches/ArduinoSketch1-0/helper_methods.h
@@ -2,4 +2,4 @@
 #include <Arduino.h>
 int move_servo_by_one_degree(Servo&servo, int final_angle);
 void writeServo(Servo &theServo, int angle);
-float getVoltageLevel(int battery_pin, int low_battery_voltage, int voltage_th);
+float getVoltageLevel(int battery_pin, int low_battery_voltage, float voltage_th);
